+++ core/frontend-devtools/src/tools/RenderTargetTools.ts
@@ -7,7 +7,7 @@
  * @module Tools
  */
 
-import { IModelApp, PrimitiveVisibility, RenderTargetDebugControl, ScreenViewport, Tool } from "@itwin/core-frontend";
+import { IModelApp, PrimitiveVisibility, type RenderTargetDebugControl, ScreenViewport, Tool } from "@itwin/core-frontend";
 import { parseToggle } from "./parseToggle";
 
 /** Executes some code against a RenderTargetDebugControl obtained from the selected viewport.
