+++ core/frontend/src/internal/render/webgl/Target.ts
@@ -27,7 +27,8 @@ import { RenderMemory } from "../../../render/RenderMemory";
 import { createEmptyRenderPlan, RenderPlan } from "../RenderPlan";
 import { PlanarClassifierMap, RenderPlanarClassifier } from "../RenderPlanarClassifier";
 import { RenderTextureDrape, TextureDrapeMap } from "../RenderTextureDrape";
-import { PrimitiveVisibility, RenderTarget, RenderTargetDebugControl } from "../../../render/RenderTarget";
+import { RenderTarget } from "../../../render/RenderTarget";
+import { PrimitiveVisibility, RenderTargetDebugControl } from "../RenderTargetDebugControl";
 import { ScreenSpaceEffectContext } from "../../../render/ScreenSpaceEffectBuilder";
 import { Scene } from "../../../render/Scene";
 import { QueryTileFeaturesOptions, QueryVisibleFeaturesCallback } from "../../../render/VisibleFeature";
